Test refactor, so not all of the tests are multiple times. If the the test is independed of the FileFormat/Execution engine then only run ti on every Catalog
